backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
"FastAPI",
"uvicorn",
"dotenv",
"openai",
"pydantic-settings",
"selenium",
"bs4"]
